Set up preview content for homepage
This creates the 'api/preview' route, where we take the request and response object, and set a cookie to tell the browser we want to be in preview mode. The route redirects the user back to the page they wanted to see the preview on. For example, 'api/preview?route=/' will take the user to the preview version of the home page. This also tells the home page to render different content depending on ctx.preview value. Lastly, this also adds a handler to clear preview mode (can cause a lot of confusion if we don't clear it), on the route 'api/clear-preview'.